fix: move CQRS/hadlers in separate classes
