+++ Pharmacy/Pharmacy.Infrastructure.Business/CQS/Handlers/QueriesHandlers/ProductAmount/GetAllProductAmountsHandler.cs
@@ -5,25 +5,8 @@ using Pharmacy.Infrastructure.Data;
 using Pharmacy.Infrastructure.Data.DTO;
 using Pharmacy.Infrastructure.Queries;
 
-namespace Pharmacy.Infrastructure.Handlers.ProductQueriesHanders
+namespace Pharmacy.Infrastructure.Business.CQS.Handlers.QueriesHandlers.Product
 {
-    public class GetProductAmountByIdHandler : IRequestHandler<GetProductAmountByIdQuery, ProductAmountDetailsDTO>
-    {
-        private readonly UnitOfWork _uow;
-        private readonly IMapper _mapper;
-
-        public GetProductAmountByIdHandler(UnitOfWork uow, IMapper mapper)
-        {
-            _uow = uow;
-            _mapper = mapper;
-        }
-        public async Task<ProductAmountDetailsDTO> Handle(GetProductAmountByIdQuery request, CancellationToken cancellationToken)
-        {
-            var productAmount = await _uow.ProductAmount.GetAsync(request.Id);
-
-            return _mapper.Map<ProductAmountDetailsDTO>(productAmount);
-        }
-    }
 
     public class GetAllProductAmountsHandler : IRequestHandler<GetAllProductAmountsQuery, IEnumerable<ProductAmountDTO>>
     {
