+++ Pharmacy/Pharmacy.Infrastructure.Business/CQS/Handlers/QueriesHandlers/ProductType/GetAllProductTypesHandler.cs
@@ -5,26 +5,8 @@ using Pharmacy.Infrastructure.Data;
 using Pharmacy.Infrastructure.Data.DTO;
 using Pharmacy.Infrastructure.Queries;
 
-namespace Pharmacy.Infrastructure.Handlers.ProductQueriesHanders
+namespace Pharmacy.Infrastructure.Business.CQS.Handlers.QueriesHandlers.ProductType
 {
-    public class GetProductTypeByIdHandler : IRequestHandler<GetProductTypeByIdQuery, ProductTypeDetailsDTO>
-    {
-        private readonly UnitOfWork _uow;
-        private readonly IMapper _mapper;
-
-        public GetProductTypeByIdHandler(UnitOfWork uow, IMapper mapper)
-        {
-            _uow = uow;
-            _mapper = mapper;
-        }
-        public async Task<ProductTypeDetailsDTO> Handle(GetProductTypeByIdQuery request, CancellationToken cancellationToken)
-        {
-            var productType = await _uow.ProductType.GetAsync(request.Id);
-
-            return _mapper.Map<ProductTypeDetailsDTO>(productType);
-        }
-    }
-
     public class GetAllProductTypessHandler : IRequestHandler<GetAllProductTypesQuery, IEnumerable<ProductTypeDTO>>
     {
         private readonly UnitOfWork _uow;
