+++ Pharmacy/Pharmacy.Infrastructure.Business/CQS/Handlers/QueriesHandlers/Warehouse/GetAllWarehouseHandler.cs
@@ -5,25 +5,8 @@ using Pharmacy.Infrastructure.Data;
 using Pharmacy.Infrastructure.Data.DTO;
 using Pharmacy.Infrastructure.Queries;
 
-namespace Pharmacy.Infrastructure.Handlers.ProductQueriesHanders
+namespace Pharmacy.Infrastructure.Business.CQS.Handlers.QueriesHandlers.Warehouse
 {
-    public class GetWarehouseByIdHandler : IRequestHandler<GetWarehouseByIdQuery, WarehouseDetailsDTO>
-    {
-        private readonly UnitOfWork _uow;
-        private readonly IMapper _mapper;
-        public GetWarehouseByIdHandler(UnitOfWork uow, IMapper mapper)
-        {
-            _uow = uow;
-            _mapper = mapper;
-        }
-        public async Task<WarehouseDetailsDTO> Handle(GetWarehouseByIdQuery request, CancellationToken cancellationToken)
-        {
-            var warehouse = await _uow.Warehouse.GetAsync(request.Id);
-
-            return _mapper.Map<WarehouseDetailsDTO>(warehouse);
-        }
-    }
-
     public class GetAllWarehouseHandler : IRequestHandler<GetAllWarehousesQuery, IEnumerable<WarehouseDTO>>
     {
         private readonly UnitOfWork _uow;
