+++ Pharmacy/Pharmacy.Infrastructure.Business/CQS/Handlers/QueriesHandlers/Warehouse/GetWarehouseByIdHandler.cs
@@ -1,0 +1,26 @@
+﻿
+using AutoMapper;
+using MediatR;
+using Pharmacy.Infrastructure.Data;
+using Pharmacy.Infrastructure.Data.DTO;
+using Pharmacy.Infrastructure.Queries;
+
+namespace Pharmacy.Infrastructure.Business.CQS.Handlers.QueriesHandlers.Warehouse
+{
+    public class GetWarehouseByIdHandler : IRequestHandler<GetWarehouseByIdQuery, WarehouseDetailsDTO>
+    {
+        private readonly UnitOfWork _uow;
+        private readonly IMapper _mapper;
+        public GetWarehouseByIdHandler(UnitOfWork uow, IMapper mapper)
+        {
+            _uow = uow;
+            _mapper = mapper;
+        }
+        public async Task<WarehouseDetailsDTO> Handle(GetWarehouseByIdQuery request, CancellationToken cancellationToken)
+        {
+            var warehouse = await _uow.Warehouse.GetAsync(request.Id);
+
+            return _mapper.Map<WarehouseDetailsDTO>(warehouse);
+        }
+    }
+}
